requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.dev -o requirements.txt
-e /home/richard/Projects/lbrc_flask
    # via -r requirements.dev
alembic==1.14.1
    # via lbrc-flask
amqp==5.3.1
    # via kombu
argon2-cffi==23.1.0
    # via flask-security-too
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via lbrc-flask
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.3.0
    # via
    #   flask-security-too
    #   lbrc-flask
beautifulsoup4==4.13.3
    # via -r requirements.dev
billiard==4.2.1
    # via celery
bleach==6.2.0
    # via flask-security-too
blinker==1.9.0
    # via
    #   flask
    #   flask-mail
    #   flask-principal
brotli==1.1.0
    # via fonttools
cairocffi==1.7.1
    # via cairosvg
cairosvg==2.7.1
    # via lbrc-flask
celery==5.4.0
    # via lbrc-flask
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cairocffi
    #   weasyprint
chardet==5.2.0
    # via lbrc-flask
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cssselect2==0.7.0
    # via
    #   cairosvg
    #   weasyprint
defusedxml==0.7.1
    # via cairosvg
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   flask-security-too
    #   lbrc-flask
et-xmlfile==2.0.0
    # via openpyxl
faker==36.1.1
    # via -r requirements.dev
flask==3.1.0
    # via
    #   flask-admin
    #   flask-login
    #   flask-mail
    #   flask-mailman
    #   flask-principal
    #   flask-security-too
    #   flask-sqlalchemy
    #   flask-weasyprint
    #   flask-wtf
    #   lbrc-flask
flask-admin==1.6.1
    # via lbrc-flask
flask-login==0.6.3
    # via flask-security-too
flask-mail==0.10.0
    # via lbrc-flask
flask-mailman==1.1.1
    # via flask-security-too
flask-principal==0.4.0
    # via flask-security-too
flask-security-too==5.6.0
    # via lbrc-flask
flask-sqlalchemy==3.1.1
    # via
    #   flask-security-too
    #   lbrc-flask
flask-weasyprint==1.1.0
    # via lbrc-flask
flask-wtf==1.2.2
    # via
    #   flask-security-too
    #   lbrc-flask
fonttools==4.56.0
    # via weasyprint
greenlet==3.1.1
    # via sqlalchemy
hatchling==1.27.0
    # via lbrc-flask
idna==3.10
    # via email-validator
importlib-metadata==8.6.1
    # via pygal
importlib-resources==6.5.2
    # via flask-security-too
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.5
    # via flask
jsonschema==4.23.0
    # via lbrc-flask
jsonschema-specifications==2024.10.1
    # via jsonschema
kombu==5.4.2
    # via celery
lxml==5.3.1
    # via lbrc-flask
mako==1.3.9
    # via alembic
markdown==3.7
    # via lbrc-flask
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask-security-too
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
mdurl==0.1.2
    # via markdown-it-py
openpyxl==3.1.5
    # via lbrc-flask
packaging==24.2
    # via
    #   hatchling
    #   pytest
passlib==1.7.4
    # via flask-security-too
pathspec==0.12.1
    # via hatchling
pillow==11.1.0
    # via
    #   cairosvg
    #   weasyprint
pluggy==1.5.0
    # via
    #   hatchling
    #   pytest
prompt-toolkit==3.0.50
    # via click-repl
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.4.1
    # via python-ldap
pycparser==2.22
    # via cffi
pydyf==0.11.0
    # via weasyprint
pygal==3.0.5
    # via lbrc-flask
pygments==2.19.1
    # via rich
pymysql==1.1.1
    # via -r requirements.dev
pyphen==0.17.2
    # via weasyprint
pytest==8.3.5
    # via -r requirements.dev
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   celery
    #   lbrc-flask
python-dotenv==1.0.1
    # via lbrc-flask
python-ldap==3.4.4
    # via lbrc-flask
pytz==2025.1
    # via lbrc-flask
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rich==13.9.4
    # via lbrc-flask
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
setuptools==75.8.2
    # via lbrc-flask
six==1.17.0
    # via python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.38
    # via
    #   alembic
    #   flask-security-too
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.2
    # via flask-security-too
tinycss2==1.4.0
    # via
    #   cairosvg
    #   cssselect2
    #   weasyprint
tinyhtml5==2.0.0
    # via weasyprint
trove-classifiers==2025.2.18.16
    # via hatchling
types-python-dateutil==2.9.0.20241206
    # via arrow
typing-extensions==4.12.2
    # via
    #   alembic
    #   beautifulsoup4
    #   referencing
    #   sqlalchemy
tzdata==2025.1
    # via
    #   celery
    #   faker
    #   kombu
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
weasyprint==64.1
    # via flask-weasyprint
webencodings==0.5.1
    # via
    #   bleach
    #   cssselect2
    #   tinycss2
    #   tinyhtml5
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
wtforms==3.2.1
    # via
    #   flask-admin
    #   flask-security-too
    #   flask-wtf
    #   lbrc-flask
xlrd==2.0.1
    # via lbrc-flask
zipp==3.21.0
    # via importlib-metadata
zopfli==0.2.3.post1
    # via fonttools
